stable version now
Finished the core rewrite.
Added an allow dialog for users that are newly connecting. I added a
configuration file that is created to store session information. This is
cleaned up when the app is cleaned up. This lets me store the last
connected user and the session key so that it persists across reboots. I
didnt want to store in the reg because the reg is a pain!