Updates
When the Server is running as a local user and the UAC screen pop ups,
the window is dimed and a message appears indicating the screen cannot
be captured.
Adding more event logging options -- in the works.
Working on Sending Credentials across to elevate the process remotely so
UAC can be captured.